Temptative changes that must be reviewed.

+++ src/Signal.c
@@ -47,7 +47,8 @@ double (*PSDC)(double);
 double *(*Strain)(double,double *);
 double CosF;
 int MT;
-int MaxTime;
+//int MaxTime;
+double MaxTime;
 double Thr[3];
 char *GWPath;
