+++ src/Signal.h
@@ -6,10 +6,10 @@ extern double (*PSDC)(double);
 extern double (*Strain)(double,double *);
 extern double CosF;
 extern int MT;
-extern int MaxTime;
+//extern int MaxTime;
+extern double MaxTime;
 extern double Thr[3];
 extern char *GWPath;
-//extern double *Params;
 
 double Lisa(double x);
 double ILigo(double x);
